Fix (some) flickering when dragging graph with max speed
The problem was that we were resetting the snowflake and then waiting
about 4ms (the length of setTimeout with 0), then requesting the next
frame via requestAnimationFrame, and only then would we draw the
snowflake.

Since we were deleting the previous snowflake and then drawing it on
a (likely, later) frame, there was flickering.

This changes the reset to occur on the same frame we start drawing the
new snowflake, i.e., it is now an async reset.