Fix speed slider visuals not resetting on page back/forward
If the page was opened, the speed slider changed, another page
visited, and then the back button pressed, the speed slider would
visually appear to be the changed value, but the code would not
register this value and would use the default one. Now the slider
correctly matches the default value.