Fix flickering on fast growth speed & fix browser lag
1. The snowflake would flicker when dragging a graph handle when using
the fastest growth setting. This was because the reset code also reset
the current time to performance.now(), which is an issue because it
meant that we would calculate that we would need to perform no updates
in that same frame (since the time elapsed would be essentially
zero). This mean that we would clear the canvas on one frame, but then
not draw on it until the next one, creating annoying flickering.

2. The update callback now checks periodically if it has exceeded 90%
of the time budget for drawing a frame 60 times per second, and yields
control if it has. This is not too precise a check (it can still go
over with complicated enough snowflakes), but it seems good enough for
the time being