Fix performance problems
A couple of issues:

- Using the 'struct of arrays' approach for storing the fields of
Face and Branch in Snowflake caused performance problems. My best
guess is that because we needed most of all the properties at all
times, it simply polluted the cache.

- There is some weird interaction between arrow functions and
the webpack TerserPlugin which sometimes makes them very slow. This is
unfortunate, because `function () ...` syntax contributes to larger
code sizes. I'll use `function` from now on for high-performance parts
of the code, unless I can figure out what exactly is going on.